backend/requirements.txt

fastapi==0.111.0
uvicorn[standard]==0.30.0
pydantic==2.8.2
pydantic-settings==2.2.1
sqlalchemy==2.0.30
python-multipart==0.0.9
requests==2.32.3
qdrant-client==1.9.0
tiktoken==0.7.0
numpy==1.26.4
python-docx==1.1.2
python-pptx==0.6.23
PyPDF2==3.0.1
pytest==8.2.0
